+++ lib/header.rs
@@ -3,7 +3,7 @@ use std::fmt;
 /// Http header
 ///
 /// Has a name and a value.
-#[derive(Hash, Eq)]
+#[derive(Hash, PartialEq, Eq)]
 pub struct Header {
     /// Name of the Header
     pub name: String,
@@ -69,7 +69,7 @@ impl Header {
         T: fmt::Display,
     {
         let header = header.to_string();
-        let mut split_header = header.split(':');
+        let mut split_header = header.splitn(2, ':');
         if split_header.clone().count() != 2 {
             return None;
         }
@@ -95,12 +95,6 @@ impl fmt::Display for Header {
     /// ```
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}: {}", self.name, self.value)
-    }
-}
-
-impl PartialEq for Header {
-    fn eq(&self, other: &Header) -> bool {
-        self.name == other.name && self.value == other.value
     }
 }